Update README
This, to me makes more sense for someone first using the gem. When ordered by descending position, new items are added to the top of the list which makes less sense than the traditional addition to the bottom of a list as is often the case in first in first out lists.

This also makes move_higher and other methods a little easier to understand when first being introduced to the gem.