Use `default_scoped.unscope` to avoid Rails 6 deprecation warning
In rails 6 we get the following deprecation warning:

    DEPRECATION WARNING: Class level methods will no longer inherit
    scoping from `create` in Rails 6.1. To continue using the scoped
    relation, pass it into the block directly. To instead access the full
    set of models, as Rails 6.1 will, use `<model name>.default_scoped`.
    (called from acts_as_list_list at /path/to/acts_as_list/lib/acts_as_list/active_record/acts/list.rb:229)

when we call `find_or_create_by` on a model with `acts_as_list`. To get
rid of this deprecation warning, we can introduce a call to
`default_scoped` before the existing call to `unscope(:select, :where)`.
This method has been available since rails 4.x so we don't need to do a
version check for this and it should work across all rails versions.